Added some error handling and unit tests
To do:
Error handling
Unit tests

+++ SampLib.py
@@ -15,20 +15,34 @@ class sampLib():
         self.lib = np.asarray(self.lib)
 
     def dispSamples(self):
-        print("\n", "\u0332".join("Sample Library"))
-        for i in range(len(self.lib)):
-            print(i,':', self.lib[i])
+        if len(self.lib) == 0:
+            print("***Library contains no samples***")
+        else:
+            print("\n", "\u0332".join("Sample Library"))
+            for i in range(len(self.lib)):
+                print(i,':', self.lib[i])
 
     def getFilename(self, i):
         return self.lib[i]
 
     def addSample(self, file):
-        fullFileName = file + '.wav'
+        fullFileName = file
+
+        if not file.endswith('.wav'):
+            fullFileName = file + '.wav'
+
         self.lib = np.append(self.lib, fullFileName)
 
     def delSample(self, i):
+        filename = self.lib[i]
         #delete file from disk directory
-        os.remove(self.path + self.lib[i])
+        if self.lib.size > 0:
+            os.remove(self.path + self.lib[i])
 
-        #delete from in-memory sample library
-        self.lib = np.delete(self.lib, i)
+            #delete from in-memory sample library
+            self.lib = np.delete(self.lib, i)
+
+        else:
+            print("\n***Error: Library is empty***")
+
+        return filename
